realtime-bedrock-nova2/pyproject.toml
[project]
name = "bedrock-core"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aws-sdk-bedrock-runtime>=0.1.0,<0.2.0",
    "boto3==1.40.17",
    "mcp==1.13.1",
    "numpy==2.2.4",
    "python-dotenv==1.0.0",
    "requests==2.32.3",
    "rx==3.2.0",
    "scipy==1.15.2",
    "strands-agents==0.1.6",
    "websockets==15.0.1",
]
